backend/requirements.txt
# Core Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Data Validation
pydantic>=2.0
pydantic-settings>=2.0

# Database
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
alembic>=1.13.0

# Caching & Storage
redis>=5.0
minio>=7.2.0

# ML & Model Training
mlflow>=2.15.0
boto3>=1.34.0
torch>=2.2.0
transformers>=4.40.0
datasets>=2.19.0
scikit-learn>=1.4.0
scipy>=1.13.0
river>=0.21.0
transformers-interpret>=0.10.0

# Task Scheduling
apscheduler>=3.10.0

# Data Processing
numpy>=1.26.0
pandas>=2.2.0

# Monitoring & Logging
prometheus-client>=0.20.0
prometheus-fastapi-instrumentator>=7.0.0
wandb>=0.17.0

# File Processing (NEW)
python-docx>=1.0.0
PyPDF2>=3.0.0
python-multipart>=0.0.6

# Development Dependencies (optional)
# pytest>=8.0
# pytest-asyncio>=0.23.0
# httpx>=0.27.0
# ruff>=0.4.0
